CompSci - Allen Thoe
The coding aspect was fairly difficult, but tangible. The major obstacle we ran into was when we finished the code, but the X's and O's were not popping up. It took us about a week to scan through the code only to find one little glitch that was hindering the entire code. This is our functioning and finished product!